utils: add webrecorder.utils and move load_wr_config() into utils to avoid pywb dependency in webrecorder package init bump version to 3.5
